+++ packages/core/src/components/dropdown-menu/dropdown-menu.tsx
@@ -281,7 +281,7 @@ export class DropdownMenu {
       nextIndex = tabs.length - 1;
     }
 
-    tabs[nextIndex].focus();
+    tabs[nextIndex]?.focus();
   }
 
   escape = () => {